Implement clean and rename code
